+++ changelog.md
@@ -2,8 +2,12 @@
 
 ## 0.19.0
 
-- extract the `Filesystem` interface and thread it everywhere from `src/cli/invoke.ts` and tests
+- **break**: extract the `Filesystem` interface and
+  thread it everywhere from `src/cli/invoke.ts` and tests
   ([#171](https://github.com/feltcoop/gro/pull/171))
+- **break**: replace `src/project/gitignore.ts` helper `isGitignored`
+  with `src/fs/pathFilter.ts` helper `toPathFilter`
+  ([#172](https://github.com/feltcoop/gro/pull/172))
 
 ## 0.18.2
 
@@ -38,7 +42,7 @@
   ([#164](https://github.com/feltcoop/gro/pull/164))
 - make serve task work for production SvelteKit builds
   ([#163](https://github.com/feltcoop/gro/pull/163))
-- add `src/project/gitignore.ts` with `loadGitignoreFilter`
+- add `src/project/gitignore.ts` with `isGitignored` and `loadGitignoreFilter`
   ([#165](https://github.com/feltcoop/gro/pull/165))
 - add helper `toSvelteKitBasePath` to `src/build/sveltekit.ts`
   ([#163](https://github.com/feltcoop/gro/pull/163))
