+++ src/fs/node.ts
@@ -4,7 +4,7 @@ import fsExtra from 'fs-extra';
 import type {Filesystem} from './filesystem.js';
 import type {PathStats} from './pathData.js';
 import {sortMap, compareSimpleMapEntries} from '../utils/map.js';
-import type {PathFilter} from './pathData.js';
+import type {PathFilter} from './pathFilter.js';
 
 // This uses `CheapWatch` which probably isn't the fastest, but it works fine for now.
 // TODO should this API be changed to only include files and not directories?
