+++ src/fs/pathData.ts
@@ -14,8 +14,3 @@ export const toPathData = (id: string, stats: PathStats): PathData => {
 		isDirectory: stats.isDirectory(),
 	};
 };
-
-// This is a subset of the `cheap-watch` types designed for browser compatibility.
-export interface PathFilter {
-	(file: {path: string; stats: PathStats}): boolean;
-}
